requirements.txt
anyio==3.6.2
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==38.0.3
fastapi==0.87.0
h11==0.14.0
idna==3.4
pycparser==2.21
pydantic==1.10.2
PyJWT==2.6.0
sniffio==1.3.0
starlette==0.21.0
typing_extensions==4.4.0
uvicorn==0.20.0